Addition of CommanderCounters class
Wanted to extract the counter information, and symbol information into it's own class away from the CommanderMasterScreen class.

CommanderMasterScreen is a class that should hand the functionality of the view objects.

The CommanderCounters class will store all of the behaviour of the counters, and track those counters (health, commander damage, poison counters...)